Moved atomic intrinsics tests, added description
